Refactor EditFilterLayer component
Refactor EditFilterLayer component and GeoFilterMap comonents. Break out
EditFilterLayer into another file in the components folder of the maps
module. Add props for necessary parameters. Use state instead of ref for
rectLayer to make assignment easier through parameters since it is no
longer being set only within GeoFilterMap.
